Crate parquet_derive

Modules

parquet_field 🔒

Derive Macros

ParquetRecordReader

Derive flat, simple RecordReader implementations. Works by parsing a struct tagged with #[derive(ParquetRecordReader)] and emitting the correct writing code for each field of the struct. Column readers are generated in the order they are defined.

ParquetRecordWriter

Derive flat, simple RecordWriter implementations. Works by parsing a struct tagged with #[derive(ParquetRecordWriter)] and emitting the correct writing code for each field of the struct. Column writers are generated in the order they are defined.